实例...
1. bugfix;先存AILine再激活意识去进行区域点亮;
2. 对点亮的AILine进行”能量消耗”计算;
3. 准备对横向点亮的数据进行类比分析;

+++ SMG_NothingIsAll/Class/AIFoundation/AIData/AIStore/AILineStore.h
@@ -12,7 +12,7 @@
 
 +(AILine*) searchSinglePointers:(NSArray*)pointers;
 +(NSMutableArray*) searchPointer:(AIPointer*)pointer count:(NSInteger)count;
-+(NSMutableArray*) searchPointer:(AIPointer*)pointer energy:(NSInteger)energy;//根据"能量"以"pointer"为中心搜索;
++(NSMutableArray*) searchPointer:(AIPointer*)pointer energy:(CGFloat)energy;//根据"能量"以"pointer"为中心搜索;
 
 +(void) insert:(AIObject *)data;
